move impl of HandlerErrors to the same module as the struct...
cleanup file in case of data writing faliure and
don't modify the payload itself in process_multipart_field
