+++ src/http_api/errors.rs
@@ -1,3 +1,4 @@
+use hyper::http::StatusCode;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
@@ -16,4 +17,21 @@ pub enum HandlerError {
 
     #[error("internal server error: {0:#}")]
     InternalError(#[from] anyhow::Error),
+}
+
+impl HandlerError {
+    pub fn code(&self) -> StatusCode {
+        match self {
+            // the following ones are considered a bad request error
+            HandlerError::BadRequest(_) => StatusCode::BAD_REQUEST,
+            HandlerError::InvalidDestination(_) => StatusCode::BAD_REQUEST,
+
+            // Unauthorized errors
+            HandlerError::UnAuthorized(_) => StatusCode::UNAUTHORIZED,
+            HandlerError::InvalidSource(_, _) => StatusCode::UNAUTHORIZED,
+
+            // Internal server error
+            HandlerError::InternalError(_) => StatusCode::INTERNAL_SERVER_ERROR,
+        }
+    }
 }
